CaP/requirements.txt
openai == 0.28
numpy
astunparse
shapely
pygments
pyyaml